[development/diagnostics_remove_inc_files] Update API

+++ source/1_engine/rex_engine/src/diagnostics/logging/internal/sinks/sink.cpp
@@ -13,12 +13,12 @@ namespace rexlog
     {
         bool AbstractSink::should_log(rexlog::level::LevelEnum msgLevel) const
         {
-            return msgLevel >= m_level.load(rsl::memory_order_relaxed);
+            return (int32)msgLevel >= m_level.load(rsl::memory_order_relaxed);
         }
 
         void AbstractSink::set_level(level::LevelEnum logLevel)
         {
-            m_level.store(logLevel, rsl::memory_order_relaxed);
+            m_level.store((int32)logLevel, rsl::memory_order_relaxed);
         }
 
         rexlog::level::LevelEnum AbstractSink::level() const
